Add an option to "catch up" when registering a search index.
If catch_up is set to True when register() is called, then creating a relation
index model will index all previously existing objects.

I still need to write a test for it, and think about what to do if the
register() call changes the index fields but doesn't register a new model.